% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{cleanCounts}
\alias{cleanCounts}
\title{Filter counts matrix}
\usage{
cleanCounts(counts, min.lib.size = 1000, max.lib.size = 8000,
  min.reads = 1, min.detected = 1)
}
\arguments{
\item{counts}{Read count matrix. The rows correspond to genes, columns correspond to individual cells}

\item{min.lib.size}{Minimum number of genes detected in a cell. Cells with fewer genes will be removed (default: 1000)}

\item{max.lib.size}{Maximum number of genes detected in a cell. Cells with more genes will be removed (default: 8000)}

\item{min.reads}{Minimum number of reads per gene. Genes with fewer reads will be removed (default: 10)}

\item{min.detected}{Minimum number of cells a gene must be seen in. Genes not seen in a sufficient number of cells will be removed (default: 5)}
}
\value{
a filtered read count matrix
}
\description{
Filter counts matrix based on gene and cell requirements
}
\examples{
{
data(pbmcA)
dim(pbmcA)
mat <- cleanCounts(pbmcA)
dim(mat)
}

}
