% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{counts2cpms}
\alias{counts2cpms}
\title{Normalizes counts to CPM}
\usage{
counts2cpms(counts, pseudocount = 1)
}
\arguments{
\item{counts}{Read count matrix. The rows correspond to genes, columns correspond to individual cells}

\item{pseudocount}{Pseudocount for log transformation. (default: 1)}
}
\value{
a normalized matrix
}
\description{
Normalizes raw counts to log10 counts per million with pseudocount
}
\examples{
{
data(pbmcA)
mat <- counts2cpms(pbmcA)
}

}
