% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{fastPca}
\alias{fastPca}
\title{Derive principal components by SVD}
\usage{
fastPca(m, nPcs = 2, tol = 1e-10, scale = FALSE, center = TRUE, ...)
}
\arguments{
\item{m}{Matrix X}

\item{nPcs}{Number of principal components}

\item{tol}{irlba tol}

\item{scale}{Scale input}

\item{center}{Center input}

\item{...}{Additional parameters to pass to irlba}
}
\description{
Derive principal components by SVD
}
