% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getConfidentPreds}
\alias{getConfidentPreds}
\title{Use LDA model posteriors to retain only confident predictions}
\usage{
getConfidentPreds(posterior, t = 0.95)
}
\arguments{
\item{posterior}{Posterior probabilities from LDA}

\item{t}{Posteriors below this threshold are set to NA}
}
\description{
Use LDA model posteriors to retain only confident predictions
}
