% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getDifferentialGenes}
\alias{getDifferentialGenes}
\title{Differential expression analysis (adapted from PAGODA2)}
\usage{
getDifferentialGenes(cd, cols, verbose = TRUE)
}
\arguments{
\item{cols}{Column/cell group annotations. Will perform one vs. all differential expression analysis.}

\item{verbose}{Verbosity}

\item{counts}{A read count matrix. The rows correspond to genes, columns correspond to individual cells}
}
\description{
Differential expression analysis (adapted from PAGODA2)
}
