% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getKnnMembership}
\alias{getKnnMembership}
\title{Get approximate nearest neighbors}
\usage{
getKnnMembership(mat, k, method = igraph::cluster_walktrap, verbose = TRUE)
}
\arguments{
\item{mat}{Matrix of cells as columns. Features as rows (such as PCs).}

\item{k}{K-nearest neighbor parameter.}

\item{method}{Community detection method from igraph. (default: igraph::cluster_walktrap)}

\item{verbose}{Verbosity (default: TRUE)}
}
\value{
Vector of community annotations
}
\description{
Group cells into clusters based on approximate nearest neighbors
}
\examples{
{
data(pbmcA)
mat <- cleanCounts(pbmcA)
mat <- normalizeVariance(mat)
pcs <- getPcs(mat)
com <- getKnnMembership(pcs, k=30)
}

}
