% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getVariableGenes}
\alias{getVariableGenes}
\title{Get variable genes}
\usage{
getVariableGenes(mat, nGenes)
}
\arguments{
\item{mat}{Gene expression matrix}

\item{nGenes}{Number of genes}
}
\description{
Get variable genes
}
