% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{mergePredsList}
\alias{mergePredsList}
\title{Merge a list of group annotations}
\usage{
mergePredsList(pred.com, min.group.size = 30, t = 0.1, verbose = TRUE)
}
\arguments{
\item{pred.com}{List of group annotations}

\item{min.group.size}{Minumum group size}

\item{t}{Percentage increase in entropy that must be achieved to merge annotations}

\item{verbose}{Verbosity}
}
\description{
Merge a list of group annotations
}
