% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{modelLda.multicore}
\alias{modelLda.multicore}
\title{Linear discriminant analysis model}
\usage{
modelLda.multicore(mat, com, ncores = 10, verbose = TRUE, retest = TRUE)
}
\arguments{
\item{mat}{Expression matrix with cells as columns, transferable features such as genes as rows.}

\item{com}{Community annotations}

\item{ncores}{Number of cores for parallelization.}

\item{verbose}{Verbosity (default: TRUE)}

\item{retest}{Whether to retest model for accuracy}
}
\value{
LDA model
}
\description{
Identifies components that maximally discriminate among groups using a linear discriminant analysis model
}
\examples{
{
data(pbmcA)
mat <- cleanCounts(pbmcA)
mat <- normalizeVariance(mat)
pcs <- getPcs(mat)
com <- getKnnMembership(pcs, k=30)
model <- modelLda(mat, com)
}

}
