% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcA}
\alias{pbmcA}
\title{Frozen PBMCs (Donor A)}
\format{An object of class \code{dgTMatrix} with 13939 rows and 2896 columns.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcA
}
\description{
Frozen PBMCs (Donor A)
}
\keyword{datasets}
