% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcB}
\alias{pbmcB}
\title{Frozen PBMCs (Donor B)}
\format{An object of class \code{matrix} with 15325 rows and 7765 columns.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcB
}
\description{
Frozen PBMCs (Donor B)
}
\keyword{datasets}
