% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcC}
\alias{pbmcC}
\title{Frozen PBMCs (Donor C)}
\format{An object of class \code{dgTMatrix} with 16144 rows and 9493 columns.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcC
}
\description{
Frozen PBMCs (Donor C)
}
\keyword{datasets}
