% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{referenceAnnot}
\alias{referenceAnnot}
\title{Annotations for sorted CD14+ Monocytes CD19+ B Cells, CD34+ Cells, CD4+ Helper T Cells, CD4+/CD25+ Regulatory T Cells, CD4+/CD45RA+/CD25- Naive T cells, CD4+/CD45RO+ Memory T Cells, CD56+ Natural Killer Cells, CD8+ Cytotoxic T cells, CD8+/CD45RA+ Naive Cytotoxic T Cells}
\format{An object of class \code{character} of length 2140.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
referenceAnnot
}
\description{
Annotations for sorted CD14+ Monocytes CD19+ B Cells, CD34+ Cells, CD4+ Helper T Cells, CD4+/CD25+ Regulatory T Cells, CD4+/CD45RA+/CD25- Naive T cells, CD4+/CD45RO+ Memory T Cells, CD56+ Natural Killer Cells, CD8+ Cytotoxic T cells, CD8+/CD45RA+ Naive Cytotoxic T Cells
}
\keyword{datasets}
