% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{clusterBasedBatchCorrect}
\alias{clusterBasedBatchCorrect}
\title{Batch correct within identified groups}
\usage{
clusterBasedBatchCorrect(lds.all, batch, com.final, min.group.size = 10)
}
\description{
Batch correct within identified groups
}
