% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getPcs}
\alias{getPcs}
\title{Dimensionality reduction by PCA}
\usage{
getPcs(mat, nGenes = min(nrow(mat), 1000), nPcs = 100, ...)
}
\arguments{
\item{mat}{Variance normalized gene expression matrix.}

\item{nGenes}{Number of most variable genes. (default: 1000)}

\item{nPcs}{Number of principal components. (default: 100)}
}
\value{
Matrix with columns as cells and rows as principal component eigenvectors.
}
\description{
Dimensionality reduction using PCA by computing principal components using highly variable genes
}
\examples{
{
data(pbmcA)
mat <- cleanCounts(pbmcA)
mat <- normalizeVariance(mat)
pcs <- getPcs(mat)
}

}
