% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{getStableClusters}
\alias{getStableClusters}
\title{Iterative merging of clusters along tree until stable}
\usage{
getStableClusters(cd, com, matnorm, z.threshold = 3,
  hclust.method = "ward.D", min.group.size = 10, min.diff.genes = nrow(cd)
  * 0.005, max.iter = 10, verbose = TRUE, plot = FALSE)
}
\arguments{
\item{cd}{Counts matrix for differential expression analysis. Rows are genes. Columns are cells.}

\item{com}{Community/group annotations for cells}

\item{matnorm}{Normalized gene expression matrix for building group relationship tree. Rows are genes. Columns are cells.}

\item{z.threshold}{Z-score threshold for identifying significantly differentially expressed genes}

\item{hclust.method}{Hierarchical clustering method used to construct relationship tree}

\item{min.group.size}{Minimum group size for stable cluster}

\item{min.diff.genes}{Minimum number of significantly differentially expressed genes that must be identified or else groups will be merged}

\item{max.iter}{Maximum number of iterations. Will end earlier if convergence reached.}

\item{verbose}{Verbosity}

\item{plot}{Whether to plot intermediate plots (hierarchical clustering dendrograms)}
}
\description{
Iterative merging of clusters along tree until stable
}
