% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan_object.R
\docType{data}
\name{Mudan}
\alias{Mudan}
\title{R6 object for storing all relevant output from MUDAN analysis in single entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Mudan
}
\description{
R6 object for storing all relevant output from MUDAN analysis in single entity
}
\examples{
{
library(MUDAN)
data(pbmcA)
myMudanObject <- Mudan$new("PBMC", pbmcA)
myMudanObject$libSizeNormalize()
myMudanObject$varianceNormalize(plot=TRUE)
myMudanObject$dimensionalityReduction(maxit=1000)
myMudanObject$communityDetection(communityName='Infomap', communityMethod=igraph::cluster_infomap, k=10, minSize=10)
myMudanObject$communityDetection(communityName='Walktrap', communityMethod=igraph::cluster_walktrap, k=10, minSize=10)
myMudanObject$modelCommunity()
myMudanObject$getMudanEmbedding(plot=FALSE)
myMudanObject$getStandardEmbedding(plot=FALSE)
par(mfrow=c(2,2))
myMudanObject$plot(reductionType='pcs', communityName='Infomap', embeddingType='MUDAN')
myMudanObject$plot(reductionType='pcs', communityName='Walktrap', embeddingType='MUDAN')
myMudanObject$plot(reductionType='pcs', communityName='Infomap', embeddingType='PCA')
myMudanObject$plot(reductionType='pcs', communityName='Walktrap', embeddingType='PCA')
}

}
\keyword{datasets}
