% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan_object.R
\docType{data}
\name{Mudan}
\alias{Mudan}
\title{R6 object for storing all relevant output from MUDAN analysis in single entity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Mudan
}
\description{
R6 object for storing all relevant output from MUDAN analysis in single entity
}
\examples{
{
library(MUDAN)
data(pbmcA)
cd <- pbmcA[, 1:500]
myMudanObject <- Mudan$new("PBMCA", cd)
myMudanObject$normalizeCounts()
myMudanObject$normalizeVariance(plot=TRUE)
myMudanObject$getPcs(nPcs=30, maxit=1000)
myMudanObject$getComMembership(communityName='Infomap',
     communityMethod=igraph::cluster_infomap, k=30)
myMudanObject$getStableClusters(communityName='Infomap')
myMudanObject$modelCommunity(communityName='Infomap')
myMudanObject$getMudanEmbedding()
myMudanObject$getStandardEmbedding()
par(mfrow=c(1,2))
myMudanObject$plotEmbedding(communityName='Infomap', embeddingType='PCA',
     xlab=NULL, ylab=NULL, main='Standard')
myMudanObject$plotEmbedding(communityName='Infomap', embeddingType='MUDAN',
     xlab=NULL, ylab=NULL, main='MUDAN')
}

}
\keyword{datasets}
