% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcB}
\alias{pbmcB}
\title{Frozen PBMCs (Donor B)}
\format{A sparse matrix with 15325 genes (rows) and 7765 cells (columns).}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcB
}
\description{
Frozen PBMCs (Donor B)
}
\keyword{datasets}
