% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcC}
\alias{pbmcC}
\title{Frozen PBMCs (Donor C)}
\format{A sparse matrix with 16144 genes (rows) and 9493 cells (columns).}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
pbmcC
}
\description{
Frozen PBMCs (Donor C)
}
\keyword{datasets}
