% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{plotEmbedding}
\alias{plotEmbedding}
\title{Plot 2D embedding}
\usage{
plotEmbedding(emb, groups = NULL, colors = NULL, do.par = TRUE,
  cex = 0.6, alpha = 0.4, gradientPalette = NULL, zlim = NULL, s = 1,
  v = 0.8, min.group.size = 1, show.legend = FALSE,
  mark.clusters = FALSE, mark.cluster.cex = 2, shuffle.colors = F,
  legend.x = "topright", gradient.range.quantile = 0.95, verbose = TRUE,
  unclassified.cell.color = "gray70", group.level.colors = NULL, ...)
}
\description{
Plot 2D embedding
}
