% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{predictLds}
\alias{predictLds}
\title{Predict LD embedding for new dataset given old model and gene scale factor}
\usage{
predictLds(mat, model, gsf, verbose = TRUE)
}
\description{
Predict LD embedding for new dataset given old model and gene scale factor
}
