% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudan.R
\name{tsneLda}
\alias{tsneLda}
\title{Run tSNE on LDs from model}
\usage{
tsneLda(mat, model, perplexity = 30, verbose = TRUE, plot = TRUE,
  do.par = TRUE, ncores = 10, details = FALSE, ...)
}
\description{
Run tSNE on LDs from model
}
